break out social media icons into its own component
