+++ src/components/presenters/generic/card.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import './card.scss'
+import SocialMediaIconContainer from '../social-media-icon-container'
 import capitolSvg from '../../../assets/capitol.svg'
 import Img from 'react-image'
 
@@ -17,35 +18,6 @@ export default ({ person }) => {
     )
   }
 
-  const buildSocialMedia = () => {
-    const socialMediaList = person.socialMedia
-
-    if (!socialMediaList) {
-      return
-    }
-    return socialMediaList.map((socialMedia, index) => {
-      let hrefValue
-
-      switch(socialMedia.site) {
-        case "GooglePlus":
-          hrefValue = `https://plus.google.com/${socialMedia.handle}`
-          return <a href={hrefValue} key={index} target="_blank"><i className='fab fa-google-plus-square' /></a>
-        case "Twitter":
-          hrefValue = `https://twitter.com/${socialMedia.handle}`
-          return <a href={hrefValue} key={index} target="_blank"><i className='fab fa-twitter-square' /></a>
-        case "Facebook":
-          hrefValue = `https://facebook.com/${socialMedia.handle}`
-          return <a href={hrefValue} key={index} target="_blank"><i className='fab fa-facebook-square' /></a>
-        case "YouTube":
-          hrefValue = `https://youtube.com/${socialMedia.handle}`
-          return <a href={hrefValue} key={index} target="_blank"><i className='fab fa-youtube-square' /></a>
-        default:
-          console.log('Social media', socialMedia.name, 'not recognized')
-          return <span></span>
-      }
-    })
-  }
-
   const styledEmail =
     <div className='email'>
       <a href={`mailto:${person.email}`}>{person.email}</a>
@@ -60,7 +32,7 @@ export default ({ person }) => {
     <div className='card'>
       <figure className='image'>
         <a href={person.website || undefined}>
-          <Img src={[person.photoUrl, capitolSvg]} loader={<img src={capitolSvg} alt={person.name}/>}/>
+          <Img src={[person.photoUrl, capitolSvg]} loader={<img src={capitolSvg} alt={person.name} target='_blank'/>}/>
         </a>
       </figure>
       <figcaption className='caption'>
@@ -72,9 +44,7 @@ export default ({ person }) => {
           { person.phone && styledPhone }
           { person.address && buildAddress() }
         </div>
-        <div className='social-media'>
-          {buildSocialMedia()}
-        </div>
+        <SocialMediaIconContainer socialMedia={person.socialMedia} />
       </figcaption>
     </div>
   )
