+++ src/components/presenters/generic/card.test.js
@@ -1,5 +1,6 @@
 import Img from 'react-image'
 import Card from './card'
+import SocialMediaIconContainer from '../social-media-icon-container'
 import capitolSvg from '../../../assets/capitol.svg'
 
 describe('Card', () => {
@@ -16,7 +17,8 @@ describe('Card', () => {
         city: 'Lima',
         state: 'some state',
         zip: '22222'
-      }
+      },
+      socialMedia: ['facebook', 'youtube']
     }
 
     subject = shallow(<Card person={person} />)
@@ -31,7 +33,7 @@ describe('Card', () => {
   })
 
   it('displays the capitol while the person image is loading', () => {
-    expect(subject.find(Img).props().loader).toEqual(<img alt="Pedro P. Pedrozo" src="capitol.svg" />)
+    expect(subject.find(Img).props().loader).toEqual(<img alt="Pedro P. Pedrozo" src="capitol.svg" target="_blank"/>)
   })
 
   it('has the name of the person', () => {
@@ -50,42 +52,8 @@ describe('Card', () => {
     expect(subject.find('.address-lines').text()).toEqual('The WhiteHouse')
   })
 
-  describe('soical media', () => {
-    beforeEach(() => {
-      person = {
-        socialMedia: [
-          {site: 'GooglePlus', handle: 'googleHandle'},
-          {site: 'Twitter', handle: 'twitterHandle'},
-          {site: 'Facebook', handle: 'facebookHandle'},
-          {site: 'YouTube', handle: 'youtubeHandle'},
-        ]
-      }
 
-      subject = shallow(<Card person={person} />)
-    })
-
-    it('opens GooglePlus links in new tab', () => {
-        const googlePlusLink = subject.find({href: `https://plus.google.com/googleHandle`})
-
-        expect(googlePlusLink.props().target).toEqual('_blank')
-    })
-
-    it('opens Twitter links in new tab', () => {
-        const twitterLink = subject.find({href: `https://twitter.com/twitterHandle`})
-
-        expect(twitterLink.props().target).toEqual('_blank')
-    })
-
-    it('opens Facebook links in new tab', () => {
-        const facebookLink = subject.find({href: `https://facebook.com/facebookHandle`})
-
-        expect(facebookLink.props().target).toEqual('_blank')
-    })
-
-    it('opens YouTube links in new tab', () => {
-      const youtubeLink = subject.find({href: `https://youtube.com/youtubeHandle`})
-
-        expect(youtubeLink.props().target).toEqual('_blank')
-    })
+  it('renders social media links container', () => {
+    expect(subject.find(SocialMediaIconContainer).props().socialMedia).toEqual(['facebook', 'youtube'])
   })
 })
